Ignore E523 (not allowed here) error when UI blocks (#68)
fidget.nvim will raise errors when some plugins like clever-f are
used: when the 'emulated' operator is pending, a 'textlock' is held
but the event loop will still be running and trying to draw figdets.

There seems no a good way to tell whether it is safe to draw fidgets
(despite the vim.schedule deferral); surprisingly, an easy workaround
is to simply ignore the E523 error.

When it comes to killing and closing a fidget, we can also ignore
the E523 error but make sure deallocating fidget instances happens
after getting rid of the buffer and the window.