Do not delete the fidget buffer if it is the current one
When exiting using `:wqa` and if there are some LSP progress
notifications on exit, for example due to synchronous format-on-write,
Neovim will crash when trying to delete the fidget buffer.

For some reason, the fidget buffer becomes the active buffer in the
current window (probably because all other buffers should be closed -
although they still appear in `nvim_list_bufs`). Attempting to delete
that buffer calls `reset_synblock` [0] and the `xfree` [1] call inside
of it crashes.

We can easily prevent this Neovim crash by only deleting the fidget
buffer when it is not the current one. Such a situation should happen
only when exiting Neovim.

[0]: https://github.com/neovim/neovim/blob/1af4bd04f9ad157edbfea30642250e854c5cb5d2/src/nvim/buffer.c#L1520
[1]: https://github.com/neovim/neovim/blob/1af4bd04f9ad157edbfea30642250e854c5cb5d2/src/nvim/syntax.c#L3022