fix: right-justify by padding with spaces in Lua, rather than using :right (#158)
Fix #156 

I originally used :right because I wanted to avoid repeated string operations in Lua. The idea was to dump the set of non-padded lines into a buffer, and let Neovim handle the right-justification in-place and all at once.

But :right has the following downsides:
- its behavior is dependent on Vim settings for tabs vs spaces
- it makes things harder to debug (compared to padding spaces in Lua)
- it probably incurs some overhead of its own that isn't insignificant, because this command was meant as an interactive command.

Worst of all, I’m not sure how to benchmark any of this, so this micro-optimization is certainly premature. There was never any evidence that just doing things in Lua led to any perceivable overhead anyway.

---------

Co-authored-by: 0xadk <0xadk@users.noreply.github.com>
Co-authored-by: j-hui <j-hui@cs.columbia.edu>