feat!: Switch to event-driven LSP progress polling by default
Event-driven polling seems to be the "right" way to address #167;
see also https://github.com/neovim/neovim/pull/26098.

Periodic polling is still supported, and useful for issues like #73, but
since Neovim's built-in ringbuf configuration is not designed for the
periodic use case, it doesn't make sense for this to be Fidget's default
behavior.