fix: trigger luarocks workflow on release
This workflow was never triggering in the first place, leading to the
issue encountered in #234.

I suspect it is because triggering on the "created" event doesn't work.
According to [documentation][release-trigger], "created" isn't triggered
when draft releases are created. Not really sure how "Release Please"
works but maybe it's creating a temporary draft before publishing?

In any case, triggering on "published" seems to be what we're actually
after in this case. Not 100% sure this fix will work but we shall see
next time I bump the version.

[release-trigger]: https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#release