feat: ignore messages via to filter functions
Previously, messages could only be ignored according to the name of the
LSP server they come from. (Prior to the legacy rewrite, this feature
was part of the user-specified render function. Kind of ugly how that
worked.)

This feature allows messages to be ignored according to logic in
arbitrary user-specified functions.

The slightly yucky thing about this design is that the "shape" of the
message object is not really well-specified or documented. For example,
it is not obvious why we need to check `msg.title` rather than anything
else. The example should be good enough though, at least for now.

Addresses #249