fix: don't use client IDs from LspProgressUpdate
Fixes #253

The error was introduced in #252, which assumed that the
LspProgressUpdate User event (<=0.9.4) carries the LSP client ID
like the LspProgress event (>=10.0.0). It does not.

This commits reverts to the original behavior of only polling active
clients for pre-v10 Neovim versions (more accurately, for Neovim builds
that lack the lsp.status/LspProgress features, which were released
sometime between 0.9.4 and 10.0.0). It also moves client_ids table into
the progress/lsp.lua module, because it is specific to a particular
Neovim version. I also added some documentation.