feat: render notifications with extmarks
Prior to this change, we were manually constructing an array of strings
which represented each line in the notifications buffer, and then
separately maintaining an array of highlights to apply to those lines.
Not only is this cumbersome and error-prone, it also incurs the cost of
a bunch of repeated string allocations.

Instead, this change sets the buffer full of empty lines, and then
renders each line as an ext_mark with some attached virt_text which
accepts a sort of highlighted-twine data structure, i.e., a list of
(text, highlight[]) tuples which are horizontally concatenated.

In other words, we let ext_marks do the heavy lifting of:

-       horizontally concatenation
-       highlighting
-       right-justification

The data structure may also open doors to caching and finer-grained
highlights, though neither of these are implemented yet.

Note that this commit collided with 2a29bfd5f8529f1e1ad3868a4f820de972189c90
so the issue will regress in this commit. I will address that later.