fix: use strwidth() instead of strdisplaywidth()
The result of strdisplaywidth() depends on the current window and buffer

Supersedes 2a29bfd5f8529f1e1ad3868a4f820de972189c90 (from #280) which
uses nvim_buf_call() and nvim_win_call() to call strdisplaywidth() in
the context of the notification window, and delays the computation of
the view width until later. However, the result of that will still
depend on the previous state of the notification window, which might be
smaller than can accommodate the string length and affect the measured
length.